Demo compatible default datatype

+++ DeepPhysX_Torch/FC/FCConfig.py
@@ -35,7 +35,7 @@ class FCConfig(TorchNetworkConfig):
                  network_name: str = "FCNetwork",
                  which_network: int = 0,
                  save_each_epoch: bool = False,
-                 data_type: str = 'float64',
+                 data_type: str = 'float32',
                  lr: Optional[float] = None,
                  require_training_stuff: bool = True,
                  loss: Any = None,
