+++ DeepPhysX_Torch/Network/TorchNetworkConfig.py
@@ -42,7 +42,7 @@ class TorchNetworkConfig(BaseNetworkConfig):
                  network_type: str = 'TorchNetwork',
                  which_network: int = 0,
                  save_each_epoch: bool = False,
-                 data_type: str = 'float64',
+                 data_type: str = 'float32',
                  lr: Optional[float] = None,
                  require_training_stuff: bool = True,
                  loss: Optional[Any] = None,
